react router navigate on login/signup/login

+++ src/routes/signOut.jsx
@@ -1,17 +1,22 @@
 import { useEffect } from 'react'
+import { useNavigate } from 'react-router-dom';
 
 export default function SignOut() {
+  const navigate = useNavigate();
 
   useEffect(() => {
-    localStorage.clear();
+    if (localStorage.getItem('token')) {
+      localStorage.clear();
+    }
+    return navigate('/')
   }, [])
 
 
   return (
     <div className="container">
-      <h2>You are Logged out</h2>
+      <h2>Logging Out</h2>
       <div>
-        Thanks.
+        Please Wait...
       </div>
     </div>
   );
